requirements.txt
google-cloud-storage
google-cloud-documentai
google-genai
python-dotenv
